refactor(floorplan): parse how the top args connect to physical ports
